Port gemm_fpga
